requirements.txt
# Servidor.py (dashboard / API)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Anomalias tf.py (ML)
numpy>=1.23.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
joblib>=1.3.0
tensorflow>=2.15.0
